dreamweaver-backend/pyproject.toml
[project]
name = "dreamweaver-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
